Изменена минимально поддерживаемая версия iOS, добавлено скрытие клавиатуры после ввода названия трекера, добавлен показ заглушки при отсутствии трекера, добавлено появление галочки на кнопке отметки трекера выполненным, добавлена возможность создать нерегулярное событие
