+++ TrackerProject/Assets.xcassets/NotFoundTrackerPlaceholder.imageset/Contents.json
@@ -1,15 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "done_button.svg",
+      "filename" : "2@1x.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
+      "filename" : "2@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "2@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
