+++ TrackerProject/CreateHabbitViewController/EmojiCollectionViewCell.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class EmojiCollectionViewCell: UICollectionViewCell {
+final class EmojiCollectionViewCell: UICollectionViewCell {
     
     var titleLabel: UILabel = {
        let label = UILabel()
