+++ TrackerProject/Models/TrackerType.swift
@@ -1,0 +1,13 @@
+//
+//  TrackerType.swift
+//  TrackerProject
+//
+//  Created by Артём Костянко on 25.10.23.
+//
+
+import Foundation
+
+enum TrackerType {
+    case habbit
+    case irregularEvent
+}
